规划库改名字；规划设置起点、终点均为原始地图坐标。
1. 规划库名字从astars改为planners，用以兼容后续的rrt等。
2. 现在设置起点和终点的函数接口中，输入的坐标xy应为原始坐标的xy，然后利用mapinfo在规划器内部转换成实际的栅格坐标（针对astar类型的）。
因为例如rrt起点终点坐标可以不再栅格上，只需要离散值计算就可以，而栅格地图只是用来排除障碍物的，因此统一将接口改为原始坐标点。
3. 上述2中也体现了面对栅格地图的看法：栅格地图只是提供障碍物信息的载体，不是与算法绑定的。例如astar是只能依赖于栅格，因此需要起点终点就在栅格上；但是rrt这种起点终点完全可以不再栅格上。因此需要加深认识栅格地图。

+++ src/global_planning/include/global_planning/MCAstar.h
@@ -323,28 +323,28 @@ public:
     /// @param map 输入的原始地图
     /// @return 地图设置是否成功
     bool setMap(const cv::Mat & map) override;
-    /// @brief 设置规划路径的起点。以栅格坐标形式，而非行列形式。
-    /// @param x 栅格坐标系的x值
-    /// @param y 栅格坐标系的y值
+    /// @brief 设置规划路径的起点。以真实地图坐标形式，而非行列形式。
+    /// @param x 真实地图坐标系的x值
+    /// @param y 真实地图坐标系的y值
     /// @return 该点是否能够成为起点。即该点在地图内部且不在障碍物上。
-    bool setStartPoint(const int x, const int y) override;
-    /// @brief 设置规划路径的起点。以栅格坐标形式，而非行列形式。
-    /// @param p 栅格坐标系的点
+    bool setStartPoint(const double x, const double y) override;
+    /// @brief 设置规划路径的起点。以真实地图坐标形式，而非行列形式。
+    /// @param p 真实地图坐标系的点
     /// @return 该点是否能够成为起点。即该点在地图内部且不在障碍物上。
-    bool setStartPoint(const cv::Point2i p) override;
+    bool setStartPoint(const cv::Point2d p) override;
     /// @brief 设置规划路径的起点的朝向。
     /// @param yaw x轴为0，右手坐标系；单位为弧度；范围为[-pi, pi]。
     /// @return 该设置有效。规划器无设置起点朝向的功能时，返回false，否则返回true。
     bool setStartPointYaw(const double yaw) override;
-    /// @brief 设置规划路径的终点。以栅格坐标形式，而非行列形式。
-    /// @param x 栅格坐标系的x值
-    /// @param y 栅格坐标系的y值
+    /// @brief 设置规划路径的终点。以真实地图坐标形式，而非行列形式。
+    /// @param x 真实地图坐标系的x值
+    /// @param y 真实地图坐标系的y值
     /// @return 该点是否能够成为终点。即该点在地图内部且不在障碍物上。
-    bool setEndPoint(const int x, const int y) override;
-    /// @brief 设置规划路径的终点。以栅格坐标形式，而非行列形式。
-    /// @param p 栅格坐标系的点
+    bool setEndPoint(const double x, const double y) override;
+    /// @brief 设置规划路径的终点。以真实地图坐标形式，而非行列形式。
+    /// @param p 真实地图坐标系的点
     /// @return 该点是否能够成为终点。即该点在地图内部且不在障碍物上。
-    bool setEndPoint(const cv::Point2i p) override;
+    bool setEndPoint(const cv::Point2d p) override;
     /// @brief 设置规划路径的终点的朝向。
     /// @param yaw x轴为0，右手坐标系；单位为弧度；范围为[-pi, pi]。
     /// @return 该设置有效。规划器无设置终点朝向的功能时，返回false，否则返回true。
