+++ src/global_planning/include/global_planning/global_planner_interface.h
@@ -96,15 +96,15 @@ public:
         ori_y_ = ori_y;
         init_map_info_ = true;
     }
-    /// @brief 设置规划路径的起点。以栅格坐标形式，而非行列形式。
-    /// @param x 栅格坐标系的x值
-    /// @param y 栅格坐标系的y值
+    /// @brief 设置规划路径的起点。以真实地图坐标形式，而非行列形式。
+    /// @param x 真实地图坐标系的x值
+    /// @param y 真实地图坐标系的y值
     /// @return 该点是否能够成为起点。即该点在地图内部且不在障碍物上。
-    virtual bool setStartPoint(const int x, const int y) = 0;
-    /// @brief 设置规划路径的起点。以栅格坐标形式，而非行列形式。
-    /// @param p 栅格坐标系的点
+    virtual bool setStartPoint(const double x, const double y) = 0;
+    /// @brief 设置规划路径的起点。以真实地图坐标形式，而非行列形式。
+    /// @param p 真实地图坐标系的点
     /// @return 该点是否能够成为起点。即该点在地图内部且不在障碍物上。
-    virtual bool setStartPoint(const cv::Point2i p) = 0;
+    virtual bool setStartPoint(const cv::Point2d p) = 0;
     /// @brief 设置规划路径的起点的朝向。
     /// @param yaw x轴为0，右手坐标系；单位为弧度；范围为[-pi, pi]。
     /// @return 该设置有效。规划器无设置起点朝向的功能时，返回false，否则返回true。
@@ -113,15 +113,15 @@ public:
         std::cout << "[GlobalPlannerInterface]: \"setStartPointYaw\" is not implemented in this planner!" << std::endl;
         return false;
     }
-    /// @brief 设置规划路径的终点。以栅格坐标形式，而非行列形式。
-    /// @param x 栅格坐标系的x值
-    /// @param y 栅格坐标系的y值
+    /// @brief 设置规划路径的终点。以真实地图坐标形式，而非行列形式。
+    /// @param x 真实地图坐标系的x值
+    /// @param y 真实地图坐标系的y值
     /// @return 该点是否能够成为终点。即该点在地图内部且不在障碍物上。
-    virtual bool setEndPoint(const int x, const int y) = 0;
-    /// @brief 设置规划路径的终点。以栅格坐标形式，而非行列形式。
-    /// @param p 栅格坐标系的点
+    virtual bool setEndPoint(const double x, const double y) = 0;
+    /// @brief 设置规划路径的终点。以真实地图坐标形式，而非行列形式。
+    /// @param p 真实地图坐标系的点
     /// @return 该点是否能够成为终点。即该点在地图内部且不在障碍物上。
-    virtual bool setEndPoint(const cv::Point2i p) = 0;
+    virtual bool setEndPoint(const cv::Point2d p) = 0;
     /// @brief 设置规划路径的终点的朝向。
     /// @param yaw x轴为0，右手坐标系；单位为弧度；范围为[-pi, pi]。
     /// @return 该设置有效。规划器无设置终点朝向的功能时，返回false，否则返回true。
